rating changed back to integer from [1-5]

+++ src/main/java/com/mako/wawel/entity/movies/Review.java
@@ -18,7 +18,7 @@ public class Review {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Double rating;
+    private Integer rating;
 
     @Column(name = "review_text")
     private String reviewText;
